Phase 3a: Extract shared modules from ProBillSplitter
- Extract PERSON_COLORS, SPLIT_METHOD_OPTIONS, formatCurrency to lib/design-tokens.ts
- Extract SplitSimpleIcon to components/SplitSimpleIcon.tsx
- Extract GridCell to components/GridCell.tsx
- Extract bill calculations to hooks/use-bill-calculations.ts
- Update all imports (MobileSpreadsheetView, MobileCardView)
- ProBillSplitter: 2121 → 1932 lines
