+++ components/MobileSpreadsheetView.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from "react"
 import { MobileCardView } from "./mobile/MobileCardView"
 import { MobileGridView } from "./mobile/MobileGridView"
 import { ViewToggle } from "./mobile/shared/ViewToggle"
-import { SplitSimpleIcon } from "@/components/ProBillSplitter"
+import { SplitSimpleIcon } from "@/components/SplitSimpleIcon"
 import { ReceiptScanner } from "@/components/ReceiptScanner"
 import { ShareBill } from "@/components/ShareBill"
 import { Input } from "@/components/ui/input"
