+++ components/mobile/MobileCardView.tsx
@@ -37,7 +37,7 @@ import { formatCurrency } from "@/lib/utils"
 import { copyToClipboard, generateSummaryText } from "@/lib/export"
 import { useToast } from "@/hooks/use-toast"
 import { useBillAnalytics } from "@/hooks/use-analytics"
-import { SplitSimpleIcon } from "@/components/ProBillSplitter"
+import { SplitSimpleIcon } from "@/components/SplitSimpleIcon"
 import { BillLookup } from "@/components/BillLookup"
 import { ShareBill } from "@/components/ShareBill"
 import { ReceiptScanner } from "@/components/ReceiptScanner"
